Update text size post-applying a preset, too.
